oneboard_web/requirements.txt
Django==4.2.3
psycopg2==2.9.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.3.0
drf-nested-routers==0.94.1